+++ integration_test/player_renderer_settings_test.dart
@@ -5,7 +5,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
-import 'package:local_tag_player/src/app.dart';
+import 'package:local_tag_player/src/core/playback_settings.dart';
+import 'package:local_tag_player/src/pages/library/library_page.dart';
 
 // ignore_for_file: slash_for_doc_comments